pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "contract-redliner-pro"
version = "0.1.0"
description = "Demo-ready LangGraph contract redlining system with FastAPI, Streamlit, and DOCX tracked revisions"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.29",
  "pydantic>=2.7",
  "python-multipart>=0.0.9",
  "langgraph>=0.2.0",
  "langchain-core>=0.2.0",
  "streamlit>=1.33",
  "python-docx>=1.1.0",
  "lxml>=5.2.1",
  "httpx>=0.27",
  "tenacity>=8.2",
  "orjson>=3.10",
  "jinja2>=3.1"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
